Trunk, tools, debugger, got it to compile on Mac OSX, cleanup strongly encouraged.
git-svn-id: svn://localhost/public/MoSync/trunk@1330 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ tools/debugger/cmd_exec.cpp
@@ -347,7 +347,7 @@ static bool setTempBreak(int address) {
 }
 
 static bool removeTempBreak(int address) {
-	_ASSERT(gTempBreakpoint.callback != NULL && gTempBreakpoint.address == address);
+	_ASSERT(gTempBreakpoint.callback != NULL && gTempBreakpoint.address == (uint)address);
 	StubConnection::writeCodeMemory(address, &gMemCs[address], 1, &Callback::opDone);
 	gTempBreakpoint.callback = NULL;
 	return false;
