+++ tools/debugger/helpers.cpp
@@ -128,7 +128,7 @@ bool parseArgRegName(const std::string& src, int* dst) {
 		int i;
 		int len;
 		int res = sscanf(src.c_str() + 1, "%i%n", &i, &len);
-		if(res == 1 && len+1 == src.length()) {
+		if(res == 1 && len+1 == (int)src.length()) {
 			*dst = 32 + i;
 			return true;
 		}
